Task1: add abstract class Cryptocurrency. Some refactor. (fixes #11)
