+++ src/Task1/CoinMarket.php
@@ -8,7 +8,7 @@ class CoinMarket
 
     public function addCurrency(Currency $currency): void
     {
-        array_push($this->currencies, $currency);
+        $this->currencies[] = $currency;
     }
 
     public function maxPrice(): float
@@ -16,8 +16,8 @@ class CoinMarket
         $result = 0;
 
         if(!empty($this->getCurrencies())){
-            $result = array_reduce($this->currencies,function($carry, Currency $item){
-                return max($item->getDailyPrice(),$carry);//args order is important when daily price == 0
+            $result = array_reduce($this->currencies, function($carry, Currency $item){
+                return max($item->getDailyPrice(), $carry);//args order is important when daily price == 0
             });
         }
         return $result;
